Add killer move heuristic
Attempts to search moves that previuosly were known to be good at the same depth as the current search. If a move was good, there are typically only a few moves that meet the threat and maintain equality. This greatly restricts the search tree because our opponent must meet our threat or lose.

Bench: 7153794