Add quiescence search
Quiescence search  avoids the horizion effect. That is, the fact that we have to stop searching somewhere, but some moves - like captures - radically change the evaluation. Qsearch is a way of continuing the search until every line ends in a quiet move which is much more likely to not have radical evaluation changes on the next move. This way, we're not falsely evaluating a position as good for us just because we captured a pawn with our queen, only to realize that we're losing our queen next move

Bench: 11859988