Remove illegal moves from pv
After lazy checking legality, we sometimes run into the issue that the move in the PV is illegal. This occurs when no move changes alpha and the first move in the list is illegal. Since no move changed alpha, there is no benefit with having a valid move in the TT. Therefore, it's not worth trying to update this to be a legal move. We instead add a check to make sure that the move is legal before adding it to the pv.

Bench: 13749273