Fix index out of bound when accessing killers
Fix index out of bounds error and an integer overflow error near the end of the game. When the game is at a mate in one, and we had previously searched the board at a mate-in-two position, we had an error where the search would run to the maximum depth and then try to access the killers board out of bounds. This fixes this issue and the issue of overflowing our depth on long searches

Bench: 13749273