Use arrayvec for moveordering
Redoing the previous commits. The issue was that the tests were not running in release mode, so they would overflow the stack. I've changed it so that by default in VSCode tests are run using release. I decided to not do it by default for every run of tests to make debugging easier.

Bench: 13749273